clean: 清掉 clippy 的 8 条提示 —— 全是等价改写，不涉语义
`cargo clippy --workspace` 报了 8 条（rustc 的 `dead_code` 基线不算）：

| lint | 位置 | 改法 |
|---|---|---|
| `unnecessary_cast` | `crates/env/src/wire/frompair.rs:86` | `TaskId((v1 >> 32) as usize)` → `TaskId(v1 >> 32)`（`v1` 本就是 `usize`） |
| `question_mark` | `crates/protocol/src/console/server.rs:378` | `let Some(r) = … else { return None }` → `let r = …?` |
| `question_mark` | `scheduler/core/hart.rs:304` | 同上（`advance()` 返 `Option<usize>`） |
| `needless_range_loop` | `lock/depend.rs:92` | `for i in 0..held.len { held.slots[i]… }` → `for slot in &held.slots[..held.len]` |
| `type_complexity` ×2 | `wait/holder.rs:45`、`wait/site.rs:161` | 抽 `type HolderTable` / `type Shard`（紧邻的 `sites()` 一并对齐） |
| `doc_lazy_continuation` ×2 | `wait/site.rs:222` | 列表后的段落补一个空 doc 行，别让 rustdoc 把它读成列表项的续行 |

判据

| | 值 |
|---|---|
| clippy（默认档） | **0 条** clippy 提示；只剩 7 条 rustc `dead_code`（audit 门控类型，`cargo check` 里同样存在，是既有基线） |
| clippy（audit 档） | 只剩 1 条 `NotInitialized`（同上，属基线） |
| 门（full） | **5/5**，rc 0（默认 ×3 / audit / harden） |
| `cargo fmt --all --check` | 0 处 diff |
| 默认档告警 | kernel 7 + manifest 4 —— 与基线相同 |

自伤一处并修掉：抽类型别名时把它插在了原函数的文档块**之后**，那段文档于是改挂在别名上
（`holder.rs` 与 `site.rs` 各一处）——已把别名连同自己的一行文档挪到文档块之前。

+++ crates/env/src/wire/frompair.rs
@@ -83,7 +83,7 @@ impl FromPair for (PieToken, crate::permission::Permission) {
 impl FromPair for (PieToken, crate::permission::Permission, TaskId) {
     fn from_pair(v0: usize, v1: usize) -> Self {
         let permission = crate::permission::Permission::from_bits_truncate(v1 as u32);
-        let vestor = TaskId((v1 >> 32) as usize);
+        let vestor = TaskId(v1 >> 32);
         (PieToken(v0), permission, vestor)
     }
 }
